+++ CHANGELOG.md
@@ -4,7 +4,8 @@ New Features:
 Fixes:
 * Stack overflow error provoked by scope resolution operator
 * Compilation with invalid indentation
-* Print of If statement in AST form
+* Print of `If` statement in AST form
 * Putting a comment in the middle don't compile
 * Line with only white spaces don't compile
 * Comparison of non-integers
+* Compiling function with all `Return` statements inside of an `If` statements